exp: ZITboost 실험 노트북 + 모듈 버그 수정 + 전처리 개선
- zitboost_experiment.ipynb: 후처리 비활성화(mean 고정, pi/zc off), 테스트 실행 완료
- zi_tweedie.py: 마이너 수정
- outlier.py, aggregation.py, meta_features.py, data.py: 전처리 안정성 개선
- 논문요약.md: 정리

+++ 3_modeling/modules/zi_tweedie.py
@@ -302,7 +302,7 @@ class ZITboostRegressor(BaseEstimator, RegressorMixin):
         -------
         self
         """
-        X = np.asarray(X, dtype=np.float32)
+        X = np.asarray(X, dtype=np.float64)
         y = np.asarray(y, dtype=np.float64).ravel()
 
         # 초기화
@@ -392,7 +392,7 @@ class ZITboostRegressor(BaseEstimator, RegressorMixin):
         if not hasattr(self, "fitted_"):
             raise ValueError("Model not fitted. Call fit() first.")
 
-        X = np.asarray(X, dtype=np.float32)
+        X = np.asarray(X, dtype=np.float64)
         pi = self.lgb_pi_.predict(X)  # cross_entropy: 확률 직접 반환
         pi = np.clip(pi, 1e-8, 1 - 1e-8)
         mu = self.lgb_mu_.predict(X)
